+++ projects/ngx-mat-datetime-picker/src/lib/core/ngx-mat-native-date-formats.ts
@@ -10,7 +10,7 @@ import { MatDateFormats } from '@angular/material/core';
 
 const DEFAULT_DATE_INPUT = {
   year: 'numeric', month: 'numeric', day: 'numeric',
-  hour12: false, hour: "2-digit", minute: "2-digit", second: "2-digit"
+  hour12: true, hour: "2-digit", minute: "2-digit", second: "2-digit"
 }
 
 export const NGX_MAT_NATIVE_DATE_FORMATS: MatDateFormats = {